+++ packages/xrpl/HISTORY.md
@@ -6,6 +6,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Support for `Single Asset Vault` (XLS-65)
+* Add new amount types: `XRPLNumber` and `ClawbackAmount`
 
 ### Fixed
 * Fix `AccountRoot` ledger object to correctly parse `FirstNFTokenSequence` field.