+++ packages/xrpl/HISTORY.md
@@ -4,6 +4,11 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ## Unreleased
 
+### Fixed
+* Fix `AccountRoot` ledger object to correctly parse `FirstNFTokenSequence` field.
+
+## 4.3.0 (2025-6-09)
+
 ### Added
 * Support for `NFTokenMintOffer` (XLS-52)
 * Add Github Actions step to auto-generate documentation